backend/requirements.txt
flask==1.1.2
flask-restplus==0.13.0
sqlalchemy==1.3.23
onnxruntime==1.4.0
Shapely==1.7.1
Werkzeug==0.16.1
PyMySQL==1.0.2
Pillow==8.1.0
pyclipper==1.2.1
torch==1.7.1
torchvision==0.8.2
Werkzeug==0.16.1
opencv-python